+++ Assets/Scripts/Web/AppUser/PhoneRequest.cs
@@ -2,8 +2,8 @@
 
 public class PhoneRequest
 {
-    public int Id { get; set; }
-    public int PhoneCountryId { get; set; }
+    public long Id { get; set; }
+    public long PhoneCountryId { get; set; }
     public String Phone { get; set; }
 
     public PhoneRequest()
@@ -11,7 +11,7 @@ public class PhoneRequest
 
     }
 
-    public PhoneRequest(int id, int phoneCountryId, String phone)
+    public PhoneRequest(long id, long phoneCountryId, String phone)
     {
         Id = id;
         PhoneCountryId = phoneCountryId;
