+++ Assets/Scripts/Web/Referred/ReferredWebOperations.cs
@@ -40,7 +40,7 @@ public class ReferredFullsGetOperation : HttpOperation
 public class ReferredGetOperation : HttpOperation
 {
     [HttpQueryString]
-    public int appUserId;
+    public long appUserId;
 
     [HttpResponseJsonBody]
     public List<Referred> referreds;
@@ -102,5 +102,5 @@ public class ReferredPutOperation : HttpOperation
     public Referred referred;
 
     [HttpResponseJsonBody]
-    public int referredlId;
+    public long referredlId;
 }
